Add own envelope item header tests
