+++ Tests/SentryTests/Protocol/SentryEnvelopeTests.swift
@@ -310,39 +310,7 @@ class SentryEnvelopeTests: XCTestCase {
         XCTAssertEqual(data2.count, 3)
     }
 
-    func test_SentryEnvelopeItemHeaderSerialization_DefaultInit() {
-        let header = SentryEnvelopeItemHeader(type: "SomeType", length: 10)
-
-        let data = header.serialize()
-        XCTAssertEqual(data.count, 2)
-        XCTAssertEqual(data.count, 2)
-        XCTAssertEqual(data["type"] as? String, "SomeType")
-        XCTAssertEqual(data["length"] as? Int, 10)
-        XCTAssertNil(data["filename"])
-        XCTAssertNil(data["content_type"])
-    }
     
-    func test_SentryEnvelopeItemHeaderSerialization_WithoutFileName() {
-        let header = SentryEnvelopeItemHeader(type: "SomeType", length: 10, contentType: "text/html")
-
-        let data = header.serialize()
-        XCTAssertEqual(data["type"] as? String, "SomeType")
-        XCTAssertEqual(data["length"] as? Int, 10)
-        XCTAssertNil(data["filename"])
-        XCTAssertEqual(data["content_type"] as? String, "text/html")
-        XCTAssertEqual(data.count, 3)
-    }
-    
-    func test_SentryEnvelopeItemHeaderSerialization_AllParameters() {
-        let header = SentryEnvelopeItemHeader(type: "SomeType", length: 10, filenname: "SomeFileName", contentType: "text/html")
-        
-        let data = header.serialize()
-        XCTAssertEqual(data["type"] as? String, "SomeType")
-        XCTAssertEqual(data["length"] as? Int, 10)
-        XCTAssertEqual(data["filename"] as? String, "SomeFileName")
-        XCTAssertEqual(data["content_type"] as? String, "text/html")
-        XCTAssertEqual(data.count, 4)
-    }
     
     func testInitWithDataAttachment_MaxAttachmentSize() {
         let attachmentTooBig = Attachment(data: fixture.dataTooBig, filename: "")
